requirements_final.txt
torch>=2.0.0
torchvision
pandas
numpy
matplotlib
seaborn
streamlit
plotly
scikit-learn
Pillow
scipy
paramiko
tqdm
requests
